pyproject.toml
[project]
name = "ai-twitter-agent_team"
version = "0.1.0"
description = "Une équipe d'agents IA qui manage un compte Twitter"
authors = [
    {name = "zedems01", email = "nguessie.achille@gmail.com"},
]
dependencies = [
    "tweepy",
    "streamlit",
    "pydantic",
    "python-dotenv",
    "openai",
    "duckduckgo-search",
    "googlesearch-python",
    "pycountry",
    "loguru",
    "python-dotenv",
    "ipykernel>=6.29.5",
    "agno>=1.1.7",
    "firecrawl>=1.12.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
